Classes/lecture5 tasmia (#36)
* ProfileResource added and  added security check to all the APIs

* API for Timezone,Country,University implemented

* Modifications are done

+++ src/main/java/co/getmehired/getmehired/rest/FileResource.java
@@ -12,10 +12,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.google.firebase.auth.FirebaseAuth;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.ByteArrayOutputStream;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
 
 /**
  * Created by Dell on 09-Jul-19.
@@ -28,15 +31,25 @@ public class FileResource {
 	private FileService fileService;
 
 	@PutMapping("/api/files/{fileType}")
-	public FileMeta upload(@RequestParam("file") @RequestBody MultipartFile file, @PathVariable(name = "fileType") String fileType) {
+	public FileMeta upload(@RequestHeader String idToken,@RequestParam("file") @RequestBody MultipartFile file, @PathVariable(name = "fileType") String fileType) {
 
+		
+		if(!isValidUser(idToken)) {
+			return null;
+		}
+		
 		FileMeta fileMeta = fileService.uploadFile(file, fileType);
 		return fileMeta;
 
 	}
 
 	@GetMapping("/api/files/show/{fileId}")
-	public void getFile(@PathVariable(name = "fileId") String fileId, HttpServletResponse response) {
+	public FileMeta getFile(@RequestHeader String idToken,@PathVariable(name = "fileId") String fileId, HttpServletResponse response) {
+		
+		if(!isValidUser(idToken)) {
+			return null;
+		}
+		
 		FileMeta fileMeta = fileService.getFileMetaById(fileId);
 		if(fileMeta != null && GMHConstants.STATUS_VALID.equals(fileMeta.getFileStatus())) {
 			ByteArrayOutputStream baos = fileService.getFile(fileMeta.getPath());
@@ -46,12 +59,17 @@ public class FileResource {
 			}catch (Exception ex) {
 			}
 		}
-		return;
+		return fileMeta;
 	}
 
 	@GetMapping("/api/files/download/{fileId}")
-	public ResponseEntity<Resource> downloadFile(@PathVariable(name="fileId") String fileId,  HttpServletRequest request) {
+	public ResponseEntity<Resource> downloadFile(@RequestHeader String idToken,@PathVariable(name="fileId") String fileId,  HttpServletRequest request) {
 
+		if(!isValidUser(idToken)) {
+			return null;
+		}
+		
+		
 		// TODO get path from database
 		FileMeta fileMeta= fileService.getFileMetaById(fileId);
 		if(fileMeta != null && GMHConstants.STATUS_VALID.equals(fileMeta.getFileStatus())) {
@@ -80,14 +98,35 @@ public class FileResource {
 
 	// TODO update this resource - pass the file id
 	@DeleteMapping("/api/files/{fileId}")
-	public FileMeta deleteFile(@PathVariable(name="fileId") String fileId) {
+	public FileMeta deleteFile(@RequestHeader String idToken,@PathVariable(name="fileId") String fileId) {
 
+		if(!isValidUser(idToken)) {
+			return null;
+		}
+		
 		return fileService.delete(fileId);
 	}
 
 	@GetMapping("/api/files")
-	public List<FileMeta> getFiles() {
+	public List<FileMeta> getFiles(@RequestHeader String idToken) {
+		
+		if(!isValidUser(idToken)) {
+			return null;
+		}	
 		return fileService.getAllFiles();
+	}
+
+	private boolean isValidUser(String idToken) {
+		try {
+			String uid=FirebaseAuth.getInstance().verifyIdTokenAsync(idToken).get().getUid();
+			String name=FirebaseAuth.getInstance().verifyIdTokenAsync(idToken).get().getName();
+			String email=FirebaseAuth.getInstance().verifyIdTokenAsync(idToken).get().getEmail();
+
+		} catch (InterruptedException | ExecutionException e) {
+			return false;
+		}
+
+		return true;
 	}
 
 }
